// src/types/content.ts
export type NewsEntry = {
  date: string
  text: string
}

export type ResearchArea = {
  id: string
  title: string
  body: string
}

export type CvEntry = {
  title: string
  company: string
  period: string
  desc: string
}

export type EducationEntry = {
  title: string
  school: string
  period: string
  desc: string
}

export type EngagementEntry = {
  title: string
  desc: string
}

export type TalkEntry = {
  title: string
  venue: string
  date: string
  desc: string
}